+++ modules/error_manager/builtin/cpp/Gateway.cpp
@@ -28,7 +28,7 @@ const std::wstring gatewayName = L"error_manager";
 //=============================================================================
 static const nlsGateway gateway[] = {
     { "error", Nelson::ErrorManagerGateway::errorBuiltin, 0, 1 },
-    { "warning", Nelson::ErrorManagerGateway::warningBuiltin, -1, -1 },
+    { "warning", Nelson::ErrorManagerGateway::warningBuiltin, 1, -1 },
     { "lasterror", Nelson::ErrorManagerGateway::lasterrorBuiltin, 1, 1 },
     { "lastwarn", Nelson::ErrorManagerGateway::lastwarnBuiltin, 2, 2 },
 };